Only return avatar URL if avatar is present
Discord will respond with an avatar value of `nil` if the user does
not have an avatar uploaded.  Currently, this strategy will return
a URL that only has the user ID and not the avatar hash itself in this case.

This change will return a nil image in the auth hash if no avatar is returned
from discord.